Переписан процессор тотема
